Fenwick Tree for Prefix sum with updates
Fenwick Tree / Binary Indexed Tree is an space and time optimisation technique for calculating the prefix sum (when we have updates).It can be used it other scenarios too. But it's main use is that it is easy to code (require few lines of code) and takes less space than segment trees.